Don't consume output when a string-output-stream is passed explicitly.
If a caller passes in a string-output-stream they've already constructed
and we retrieve output for it, the caller won't get any of that data if
they retrieve the output themselves later (we may actually consume the
caller's data doing this).

We need to distinguish between the case where we were passed NIL as output
and constructed our own stream (need to retrieve the output ourselves) and
the case where the caller passed in their own string-output-stream
(shouldn't retrieve output). Using the same check we used to determine
whether to create a stream in the first place does the trick.